fix(migrations): 9ac9b80f 마이그를 prod head(0183) 기준으로 재배치 — FR-only prod 승격
선생님 확定(2026-07-16): FR 2건(스탠드업 core·이슈 ID 넘버링)만 prod로 먼저 승격하고
auth 마이그(develop의 0184~0193)는 develop에 남긴다. develop 원본 0194는
down_revision="0193"이라 auth 마이그가 없는 prod(head=0183)엔 직접 적용 불가 —
0194→0183a로 리비전 ID/down_revision만 바꾸고 upgrade()/downgrade() DDL 본문은
byte-identical 유지(story_number는 auth 테이블과 FK/로직 의존 0이라 안전하게 독립
적용 가능 확인함).

파일 헤더에 미래 auth 전체 prod 승격 시 정확한 divergence 해소 절차(3단계: ①develop
0184~0193 실제 실행 ②0194는 stamp만(DDL 재실행 없이 리비전 추적만 이동, 0183a와
byte-identical이라 안전) ③head 재수렴용 머지 리비전 추가)를 명시 — 그 시점 담당자가
바로 따라할 수 있게.

검증: 신선한 DB에 main head(0183)부터 전체 체인 재적용→0183a까지 정상 완주, 단일
head(`alembic heads`=['0183a']) 확인, 관련 스위트(toolset/scope/catalog/story) 134
passed.
